requirements.txt
django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2-binary==2.9.9
drf-spectacular>=0.15.1,<0.16
Pillow==10.4.0
uwsgi>=2.0.19,<2.1